Added FragmentingFlusher to abstract the fragmentation of frames.
