+++ jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/core/internal/FragmentExtension.java
@@ -18,17 +18,13 @@
 
 package org.eclipse.jetty.websocket.core.internal;
 
-import java.nio.ByteBuffer;
-import java.util.ArrayDeque;
-import java.util.Queue;
-
 import org.eclipse.jetty.util.Callback;
-import org.eclipse.jetty.util.IteratingCallback;
 import org.eclipse.jetty.util.log.Log;
 import org.eclipse.jetty.util.log.Logger;
 import org.eclipse.jetty.websocket.core.AbstractExtension;
 import org.eclipse.jetty.websocket.core.ExtensionConfig;
 import org.eclipse.jetty.websocket.core.Frame;
+import org.eclipse.jetty.websocket.core.FrameHandler;
 import org.eclipse.jetty.websocket.core.OpCode;
 import org.eclipse.jetty.websocket.core.WebSocketComponents;
 
@@ -39,9 +35,20 @@ public class FragmentExtension extends AbstractExtension
 {
     private static final Logger LOG = Log.getLogger(FragmentExtension.class);
 
-    private final Queue<FrameEntry> entries = new ArrayDeque<>();
-    private final IteratingCallback flusher = new Flusher();
-    private int maxLength;
+    private final FragmentingFlusher flusher;
+    private final FrameHandler.Configuration configuration = new FrameHandler.ConfigurationHolder();
+
+    public FragmentExtension()
+    {
+        flusher = new FragmentingFlusher(configuration)
+        {
+            @Override
+            void forwardFrame(Frame frame, Callback callback, boolean batch)
+            {
+                nextOutgoingFrame(frame, callback, batch);
+            }
+        };
+    }
 
     @Override
     public String getName()
@@ -58,154 +65,20 @@ public class FragmentExtension extends AbstractExtension
     @Override
     public void sendFrame(Frame frame, Callback callback, boolean batch)
     {
-        ByteBuffer payload = frame.getPayload();
-        int length = payload != null ? payload.remaining() : 0;
-        if (OpCode.isControlFrame(frame.getOpCode()) || maxLength <= 0 || length <= maxLength)
+        if (OpCode.isControlFrame(frame.getOpCode()))
         {
             nextOutgoingFrame(frame, callback, batch);
             return;
         }
 
-        FrameEntry entry = new FrameEntry(frame, callback, batch);
-        if (LOG.isDebugEnabled())
-            LOG.debug("Queuing {}", entry);
-        offerEntry(entry);
-        flusher.iterate();
+        flusher.sendFrame(frame, callback, batch);
     }
 
     @Override
     public void init(ExtensionConfig config, WebSocketComponents components)
     {
         super.init(config, components);
-        maxLength = config.getParameter("maxLength", -1);
-    }
-
-    private void offerEntry(FrameEntry entry)
-    {
-        synchronized (this)
-        {
-            entries.offer(entry);
-        }
-    }
-
-    private FrameEntry pollEntry()
-    {
-        synchronized (this)
-        {
-            return entries.poll();
-        }
-    }
-
-    private class Flusher extends IteratingCallback implements Callback
-    {
-        private FrameEntry current;
-        private boolean finished = true;
-
-        @Override
-        protected Action process() throws Exception
-        {
-            if (finished)
-            {
-                current = pollEntry();
-                LOG.debug("Processing {}", current);
-                if (current == null)
-                    return Action.IDLE;
-                fragment(current, true);
-            }
-            else
-            {
-                fragment(current, false);
-            }
-            return Action.SCHEDULED;
-        }
-
-        private void fragment(FrameEntry entry, boolean first)
-        {
-            Frame frame = entry.frame;
-            ByteBuffer payload = frame.getPayload();
-            int remaining = payload.remaining();
-            int length = Math.min(remaining, maxLength);
-            finished = length == remaining;
-
-            boolean continuation = (frame.getOpCode() == OpCode.CONTINUATION) || !first;
-            Frame fragment = new Frame(continuation ? OpCode.CONTINUATION : frame.getOpCode());
-            boolean fin = frame.isFin() && finished;
-            fragment.setFin(fin);
-
-            int limit = payload.limit();
-            int newLimit = payload.position() + length;
-            payload.limit(newLimit);
-            ByteBuffer payloadFragment = payload.slice();
-            payload.limit(limit);
-            fragment.setPayload(payloadFragment);
-            if (LOG.isDebugEnabled())
-                LOG.debug("Fragmented {}->{}", frame, fragment);
-            payload.position(newLimit);
-
-            nextOutgoingFrame(fragment, this, entry.batch);
-        }
-
-        @Override
-        protected void onCompleteSuccess()
-        {
-            // This IteratingCallback never completes.
-        }
-
-        @Override
-        protected void onCompleteFailure(Throwable x)
-        {
-            // This IteratingCallback never fails.
-            // The callback are those provided by WriteCallback (implemented
-            // below) and even in case of writeFailed() we call succeeded().
-        }
-
-        @Override
-        public void succeeded()
-        {
-            // Notify first then call succeeded(), otherwise
-            // write callbacks may be invoked out of order.
-            notifyCallbackSuccess(current.callback);
-            super.succeeded();
-        }
-
-        @Override
-        public void failed(Throwable cause)
-        {
-            // Notify first, the call succeeded() to drain the queue.
-            // We don't want to call failed(x) because that will put
-            // this flusher into a final state that cannot be exited,
-            // and the failure of a frame may not mean that the whole
-            // connection is now invalid.
-            notifyCallbackFailure(current.callback, cause);
-            succeeded();
-        }
-
-        private void notifyCallbackSuccess(Callback callback)
-        {
-            try
-            {
-                if (callback != null)
-                    callback.succeeded();
-            }
-            catch (Throwable x)
-            {
-                if (LOG.isDebugEnabled())
-                    LOG.debug("Exception while notifying success of callback " + callback, x);
-            }
-        }
-
-        private void notifyCallbackFailure(Callback callback, Throwable failure)
-        {
-            try
-            {
-                if (callback != null)
-                    callback.failed(failure);
-            }
-            catch (Throwable x)
-            {
-                if (LOG.isDebugEnabled())
-                    LOG.debug("Exception while notifying failure of callback " + callback, x);
-            }
-        }
+        int maxLength = config.getParameter("maxLength", -1);
+        configuration.setMaxFrameSize(maxLength);
     }
 }
